+++ src/lib/abt_did_workshop/tx/update.ex
@@ -32,6 +32,9 @@ defmodule AbtDidWorkshop.Tx.Update do
 
     "UpdateAssetTx"
     |> Helper.get_transaction_to_sign(sender, receiver)
-    |> Helper.require_signature(user_addr, beh.description)
+    |> Helper.require_signature(
+      user_addr,
+      "Pleas confirm this update by signing the transaction."
+    )
   end
 end