+++ src/lib/abt_did_workshop/utils/util.ex
@@ -12,8 +12,13 @@ defmodule AbtDidWorkshop.Util do
 
   def get_ip do
     {:ok, ip_list} = :inet.getif()
-    ips = List.first(ip_list)
-    {i1, i2, i3, i4} = elem(ips, 0)
+
+    {ip, _broadcast, _netmask} =
+      Enum.find(ip_list, fn {ip, broadcast, _netmask} ->
+        ip != {127, 0, 0, 1} and broadcast != :undefined
+      end)
+
+    {i1, i2, i3, i4} = ip
     "#{i1}.#{i2}.#{i3}.#{i4}"
   end