+++ src/lib/abt_did_workshop_web/controllers/did_auth/wallet_controller.ex
@@ -139,7 +139,7 @@ defmodule AbtDidWorkshopWeb.WalletController do
       HTTPoison.post!(url, body, [{"content-type", "application/json"}])
 
     case Jason.decode(response) do
-      {:ok, %{"appPk" => _}} ->
+      {:ok, %{"response" => "ok"}} ->
         conn
         |> put_flash(:info, "Authentication Succeeded!")
         |> redirect(to: Routes.did_path(conn, :show))